requirements.txt
openai>=1.0.0
boto3>=1.26.0
MetaTrader5>=5.0.0
pandas>=1.5.0
numpy>=1.23.0
requests>=2.28.0
python-dotenv>=0.20.0
pytest>=7.0.0
